backend/requirements.txt
Django>=4.2.0
djangorestframework>=3.14.0
django-cors-headers>=4.0.0
requests>=2.28.0
python-dotenv>=1.0.0
django-crontab>=0.7.1
pymysql>=1.1.0
django-apscheduler==0.6.2


